backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
langchain
google-generativeai
httpx
passlib[bcrypt]
python-jose
langchain_community
langchain-google-genai
pillow
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
